'Sorting Data of Files' application v.1.0.
It was programmed in python 3.10.4 (One of the latest versions).
Author: Jonathan García.
Date: June, 2022.

The application has six modules in the directory 'src':
 • main.py - Create functions to process data and display menu.
 • functions.py - General functions: retrieve files, sort records, etc.
 • menu_functions.py - Functions to print menu and get options selected
                       by the user.
 OOP (Factory Pattern)
 • file_readers.py - Interface and classes to read different kind of
                     files format.
 • file_generators.py - Interface and classes to write and format files
                        according the extension.
 • data_clients - Definition of client classes which use the Factories
                  to process and write the data.

The OOP programming paradigm was used to implement the reading and
writing of files. Using the factory pattern to be able of read the
three kinds of files: CSV, JSON and XML;and to be able of write the
data in different formats(at the moment just create TSV files). Also
this pattern permit add formats of files easily, just implementing the
respective interfaces.

The menu and other functions were coded with structured programming.
Creating instance of the client classes (ProcessData and WriteData)
to process the data.

I used PEP 8 to the structure; but the characters' maximum per line it
was fixed in 120. The comments (and docstrings) length is conserved
with a value of 72.
Also, there are type hints in the functions/methods and variables and
variables. Following the PEP 483, 484, 526...

The format of docstrings is NumPy/SciPy, following PEP 257.